requirements.txt
authy==2.1.5
automium==0.2.6
automium-web==0.1.1
Babel==2.5.1
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
Django==1.11.5
django-extra-fields==0.6
django-phonenumber-field==1.3.0
django-twilio==0.8.0
djangorestframework==3.7.0
Flask==0.12.2
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
legacy==0.1.6
MarkupSafe==1.0
olefile==0.44
phonenumberslite==8.8.3
Pillow==4.2.1
psycopg2==2.7.3.1
PyJWT==1.5.3
PySocks==1.6.7
pytz==2017.2
requests==2.18.4
simplejson==3.11.1
six==1.11.0
twilio==6.7.1
urllib3==1.22
Werkzeug==0.12.2
